When a device is unplugged select a new device
When a user unplugs a device, the old behaviour was to
still try and use this device. This would cause some weirdness
when consumers of this component tried to consume the device
id which doesn't exist anymore.

When a device is unplugged now, we just choose another device (the
first device in the list).